collapse "nsp" task into "scripts"
