+++ lib/pkg.js
@@ -105,8 +105,7 @@ function isTestScriptUnconfigured({ scripts = {} } /* : PackageJSON */) {
   const NPM_INIT_TEST = 'echo "Error: no test specified" && exit 1';
   return (
     !scripts.test ||
-    scripts.test === NPM_INIT_TEST ||
-    scripts.test === 'npm run lint'
+    scripts.test === NPM_INIT_TEST
   );
 }
