+++ lib/tasks/npm-scripts.js
@@ -18,6 +18,20 @@ const LABEL = 'npm scripts configured';
 
 /* :: import type { TaskOptions } from '../types.js' */
 
+function audit(pkg /* : Object */) {
+  // npm 6.x has built-in "audit" command
+  if (!pkg.engines || !pkg.engines.npm || intersects(pkg.engines.npm, '<6')) {
+    pkg.scripts.nsp = 'npx -q nsp check';
+    addScript(pkg, 'test', 'npm run nsp');
+  } else {
+    addScript(pkg, 'test', 'npm audit');
+
+    delete pkg.scripts.nsp;
+    removeScript(pkg, 'test', 'npm run nsp');
+  }
+  return pkg;
+}
+
 function ava(pkg) {
   const scripts = {};
   if (pkg.devDependencies.ava) {
@@ -126,6 +140,7 @@ function fn({ cwd, isFlowProject = false } /* : TaskOptions */) {
 
       // TODO: ought to consider a better place for this?
       pkg = jest(pkg);
+      pkg = audit(pkg);
       pkg.prettier = prettier(safePkg);
 
       pkg.scripts = {
